Compressor int overflow when seeking to frame(s) beyond 2GB offset

+++ dcm4che-imageio/src/main/java/org/dcm4che3/imageio/codec/Compressor.java
@@ -336,7 +336,7 @@ public class Compressor extends Decompressor implements Closeable {
         iis.setByteOrder(pixeldata.bigEndian()
                 ? ByteOrder.BIG_ENDIAN
                 : ByteOrder.LITTLE_ENDIAN);
-        iis.seek(pixeldata.offset() + imageParams.getFrameLength() * frameIndex);
+        iis.seek(pixeldata.offset() + (long)imageParams.getFrameLength() * (long)frameIndex);
         DataBuffer db = bi.getRaster().getDataBuffer();
         switch (db.getDataType()) {
         case DataBuffer.TYPE_BYTE:
